pyproject.toml
[project]
name = "langmuir"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "matplotlib>=3.10.5",
  "numpy>=2.3.2",
  "pandas>=2.3.2",
  "pyyaml>=6.0",
  "rich>=13.0",
  "scipy>=1.16.1",
  "xarray>=2025.8.0",
  "refnx",
  "tqdm>=4.67.1",
]
